add full MIT license text
By including a copy of the exact MIT license text, we make the terms of
the merfi license very clear.

Also, in the specific case of merfi's MIT license, all derivative copies
of merfi are required to contain the disclaimer text within the MIT
license (as described in the sentence "The above copyright notice and
this permission notice shall be included in all copies or substantial
portions of the Software.") So in order to ease development and
distribution of merfi, it makes sense to simply include this text in the
upstream project.