tests: test_sign_no_files with a known-empty dir
Prior to this commit, test_sign_no_files would check '', which implies
the current working directory when running the tests.

It is safer and more precise if we explicitly point the
test_sign_no_files test at a directory location that we are certain is
empty.

This eliminates spurious test failures if files named "Release" happen
to exist in merfi's source code tree.