RpmSign: make --key arg mandatory
The rpm-sign utility requires you to specify a "--key" argument, so that
it knows the exact key with which to sign bits.

merfi uses self.key, but doesn't enforce its definition on the cmdline.
So prior to this change, if a user did not pass a --key argument to
merfi, merfi would crash.