collector: use DebRepo objects
Prior to this change, RepoCollector stored a list of strs (paths).
The .debian_release_files property would simply walk the tree looking
for "Release" files, and return them all, even if they were irrelevant
for GPG signing.

With this change, RepoCollector stores a list of DebRepo objects instead
of strs. The .debian_release_files property only returns the specific
repository Release file(s) that should be GPG-signed for Apt.

Update the help text for the gpg and rpm-sign backends to clarify that
we're seeking and signing repos now, not just any Release files.