tests: return top directory from fixture
The deb_repotree fixture was going to the work of os.path.join() to
select "jewel" every time, and then we had further code in the nested
tests to undo that work (with os.path.dirname()).

Simplify this by returning the parent dir directly in the fixture, and
picking "jewel" only for the tests that need it.